Fix bugs in keyboard teleop. Separate the launch files.
ardrone_teleop/keyboard_teleop_custom.launch is (obviously) my custom launch file. It's configured for a dvorak keyboard, and shifting your left hand one key to the right (I like having my index fingers rest on 'u' and 'h' ('f' and 'j' in qwerty, these are the keys that normally have the bumps on them)).
I might add commands to get it to flip. The driver (as far as I know) currently doesn't suppor that, so I may add that in.